Deleting the sg reference only if not null (#278)

+++ osc-server/src/main/java/org/osc/core/broker/service/tasks/conformance/openstack/deploymentspec/ForceDeleteDSTask.java
@@ -62,7 +62,7 @@ public class ForceDeleteDSTask extends TransactionalTask {
         boolean osSgCanBeDeleted = DeploymentSpecEntityMgr.findDeploymentSpecsByVirtualSystemTenantAndRegion(em,
                 this.ds.getVirtualSystem(), this.ds.getTenantId(), this.ds.getRegion()).size() <= 1;
 
-        if (osSgCanBeDeleted) {
+        if (osSgCanBeDeleted && this.ds.getOsSecurityGroupReference() != null) {
             OSCEntityManager.delete(em, this.ds.getOsSecurityGroupReference());
         }
